requirements.txt
# Multimodal-Integration — runtime requirements
# Python >=3.10

# Core scientific stack
numpy>=1.24,<3
pandas>=2,<3
Pillow>=10,<12
matplotlib>=3.7,<4
tqdm>=4.66,<5
requests>=2.31,<3

# Hugging Face ecosystem
transformers>=4.41,<5
datasets>=2.20,<3
sentencepiece>=0.1.99
ftfy>=6.1.3
safetensors>=0.4

# PyTorch (install wheels appropriate for your OS/CUDA before this file)
torch>=2.2,<3
torchvision>=0.17,<1

# Testing (left here for reviewer convenience; packaging will skip it)
pytest>=8,<9